DataEngineeringProject4/requirements.txt
json5==0.9.6
kafka-python==2.0.2
pandas==2.0.3
beautifulsoup4==4.12.2
requests==2.27.1
cassandra-driver==3.28.0